@@ -1,6 +1,6 @@
 {
     "name": "lilconfig",
-    "version": "3.0.0",
+    "version": "3.1.0",
     "description": "A zero-dependency alternative to cosmiconfig",
     "main": "dist/index.js",
     "types": "dist/index.d.ts",
@@ -9,7 +9,7 @@
         "build": "tsc --declaration",
         "postbuild": "du -h ./dist/*",
         "clean": "rm -rf ./dist",
-        "test": "jest --coverage",
+        "test": "NODE_OPTIONS=--experimental-vm-modules ./node_modules/.bin/jest --coverage",
         "lint": "eslint ./src/*.ts"
     },
     "keywords": [
@@ -29,19 +29,20 @@
     "author": "antonk52",
     "license": "MIT",
     "devDependencies": {
-        "@types/jest": "^27.0.2",
-        "@types/node": "^14.18.36",
-        "@typescript-eslint/eslint-plugin": "^5.54.0",
-        "@typescript-eslint/parser": "^5.54.0",
-        "cosmiconfig": "^7.1.0",
-        "eslint": "^8.35.0",
-        "eslint-config-prettier": "^8.6.0",
-        "eslint-plugin-prettier": "^4.2.1",
-        "jest": "^27.3.1",
-        "prettier": "^2.8.4",
-        "ts-jest": "27.0.7",
-        "typescript": "4.4.4"
+        "@types/jest": "^29.5.12",
+        "@types/node": "^14.18.63",
+        "cosmiconfig": "^8.3.6",
+        "eslint": "^8.56.0",
+        "eslint-config-prettier": "^9.1.0",
+        "eslint-plugin-prettier": "^5.1.3",
+        "jest": "^29.7.0",
+        "prettier": "^3.2.5",
+        "ts-jest": "29.1.2",
+        "typescript": "^5.3.3",
+        "typescript-eslint": "^7.0.1",
+        "uvu": "^0.5.6"
     },
+    "funding": "https://github.com/sponsors/antonk52",
     "engines": {
         "node": ">=14"
     }
